Refactored calculation of COM1/2 radios to electrical.xml
The reason was that the power-btn was changed only by turning the knob.
However other scripts (like Red Griffin ATC) rely on it to get the "radio is turned on"-state.
The power-btn of com1/2 is now calculated and considered "on" when not only the knob is turned,
but also the radio is serviceable and has electrical power.