Implement Spark plug fouling
Initial implementation.
Some of the engine coughing functionality was refactored in a generic "rough engine" implementation.
There can be various sources for a rough engine, and a rough engine leads to oscillating RPM and loss of power.
The spark plug fouling implementation is nearly purely done in propulsion and engine jsbsim systems and modular:
The calculation of the underlying physics (spark plug temperature, fouling state, etc) is modelled separately
from the effect.
The fouling state is saved with the plane's savestate and restored on the next session.
The simulation was bound to the already available "spark plug icing" aircraft option which got renamed to ".../fouling".

To express the effects, a new "power loss" effect is introduced; which is implemented currently by adjusting
a new friction source. If better options get available, we can switch the effects expression method easily.

The systems are now also running on less-than-fcs fps to conserve CPU.

Fix #439