Fix damage reset with autostart
Autostart calling repair function twice (because of code modularity); the second invocation
stored the previous_damage variable by the already intermediate off-state. That made the second restore
to restore a off-value in every case shortly before the first invocation did restore on-value.

Fix is to make it safe for multiple calls by noop-ing it for subsequent calls while it's still running.