Made airplane icing optional (and on by default).
Disabling the option will remove all ice in an instant, so it's helpful for beginners or
situations where you don't want to abort a nice trip.

The option affects all icable things like wing, fuselage, pitot tube and static port.
It does not affect the fog/frost of the windshield, for that there is already a separate
option.
The setting is stored alongside the aircraft data, so disabling it is persisted
across sessions.