HI/DG: start offset refactored
The heading_indicator_dg always initializes to reference prop /orientation/heading-deg.
Without further intervention it is thus auto-calibrated to the current startup heading.

However: We want it to display the (calibrated) value from the last session, so we need to adjust the
offset to compensate the difference between the last sessions heading and the current one.
The result is that the HI keeps its calibration/display from the last session, which simulates movement
of the plane with unspinning gyro (like when doing ground handling/parking).
This forces the pilot to recalibrate the HI before takeoff when changing locations (or to enable the
autocalibration feature explicitely).

(Dedicated to Karin, my beloved wife. You always show me the correct heading,
and it is a joy to fly our journey in the same direction :X <3 )