Adjust heading indicator gyro spin up/down time.
The spin up time is given as about 3 minutes from Kellys.
The spin down time I could only find various vague mentions on the net, that for safety
the gyro needs time to spin down. That time is given in a frame from about 5 to 15 minutes.
So I tought a sensible setting would be the lower end. But in reality, this also depends
on the wear and tear of the gyros bearings.