Aux fuel pump enrichens mixture a bit
Added a dedicated function modelling the "mixture unit".
This now simulates the aux fuel pump enriching the mixture slightly (POH 7-28).
(value is guessed because of lack of footage; but the effect is considered to be small,
the chosen value is 1% richer mixture which yields a faint effect in the sim).

Also refactored the "is the fuel pump operating" switches in the fuel system file to a
dedicated switch, to make the other switches more readable.