Master switches now can be operated together
This is optional and configurable in the aircraft-GUI, but default to ON.
With ON, the Master and Avionics Master will be switched together, unless shift is pressed.
Shift mode allows access to single switch part.

The reason for this is:
 a) ALT and BAT should normally be switched together (POH 7-34)
 b) In the US, there is just a single avionics switch. Usually also here,
    you operate them together.
I did it with the shift-toggle because I found the behviour in the C172 too erratic.